courses/graphql/book-store-api/requirements.txt
alembic==1.8.1
aniso8601==7.0.0
click==8.1.3
Flask==2.2.2
Flask-GraphQL==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
promise==2.3
psycopg2-binary==2.9.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
alembic==1.8.1
aniso8601==7.0.0
astroid==2.12.9
black==22.8.0
click==8.1.3
dill==0.3.5.1
Flask==2.2.2
Flask-GraphQL==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
promise==2.3
psycopg2-binary==2.9.3
pylint==2.15.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
tomlkit==0.11.4
Werkzeug==2.2.2
wrapt==1.14.1
alembic==1.8.1
aniso8601==7.0.0
astroid==2.12.9
black==22.8.0
click==8.1.3
dill==0.3.5.1
Flask==2.2.2
Flask-GraphQL==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
promise==2.3
psycopg2-binary==2.9.3
pylint==2.15.2
python-dotenv==0.21.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
tomlkit==0.11.4
Werkzeug==2.2.2
wrapt==1.14.1
